Parser overhaul for package path canonicalization
This changes the way we process and canonicalize packages as we add them.  As a
side effect `./path` style names are now supported.

I went through the call graph for AddDir and FindTypes and found the things
that just didn't smell right, and hacked until they were better.  This is
unfortunately hard to disentangle.

Another side-effect is that AddFileForTest will not re-run all the
type-checking, so tests that use it should make sure to sort the phony files
they feed into it topologically.

Also adds lots of logging for the next poor sap who has to understand this.

This was tested against the kubernetes codebase, which found many corner cases
that I originally missed but are now handled.