Fix lint failures in the deepcopy-gen
The changes in this patch takes care of lint failures resulted due to
deepcopy generated files. The total numbers of such lint failures are
around fifty errors. There are main three types of lint errors,
1. Generated functions name have underscore in them.
2. Generated functions are missing comment.
3. Some of the generated functions have unnecessary 'else' block for a given
   'if' loop.
For example, in the generated code, few of fifity some total failures are:
/zz_generated.deepcopy.go:60:6: don't use underscores in Go names;
/zz_generated.deepcopy.go:60:1: exported function should have comment or
be unexported
/zz_generated.deepcopy.go:116:10: if block ends with a return statement,
so drop this else and outdent its block