generator: include canonical import path when generating files
Canonical import paths force `go {get,build,test,install}` to
respect the vanity import path of files. For client-go, this means
users will get an error if they place it in the wrong GOPATH path.
For example, `$GOPATH/src/github.com/kubernetes/client-go` would
fail. This import comment is ignored if the file is placed in a
"vendor" directory.

Have the generator include the import comment when generating
packages.