Optionally allow processing of *_test.go files
Currently we skip any test related go files, for a specific scenario
like verify-import-boss in upstream kubernetes, we would like to verify
the imports of test code in addition as well. So it would be great to
add a flag to specify that we need to process them as well.