Convert path to slashes before passing to build.Context.Import()
On UNIX systems, we get away with the existing code because filesystem
and package paths both use the same kind of separator.  On Windows,
sadly, that isn't the case and that causes everything to go wrong
(more specifically, we end up not being able to detect that this is a
local package, and so the output path for the generated files is
wrong).

See issue #166.