Fail if any requested type cannot be copied
When scanning types to see whether or not to generate deepcopy code,
scan all types rather than stopping at the first type that requests
deepcopy generation. If any type cannot be copied, append it to a list
of uncopyable types. If the uncopyable list is not empty, fail and print
the list.

Signed-off-by: Travis Raines <traines@konghq.com>