Increase klog log levels to avoid spammy logs
This updates klog log levels to  match the conventions in
https://github.com/kubernetes/community/blob/master/contributors/devel/sig-instrumentation/logging.md.

This is also necessary because we will be using klog.V(2) for displaying
errors in the kube-openapi API linter. Having v=2 level logs in
gengo results in very spammy logs that are not relevant to the user.

So this commit updates errors to v=3, errors that are ignored to v=4
and info logs to v=5.