make deepcopy-gen compatible with go 1.17
In Golang 1.17, the gofmt tool prefers to use `//go:build` as a build
constraint, https://pkg.go.dev/go/build#hdr-Build_Constraints. So it
seems that the `// +build` was abandoned long ago but the compatibility
was preserved until 1.17.
To ensure compatibility with Golang version older than 1.17 so from 1.16
and below, having both constraints seems to work fine.

Now deepcopy-gen will generate a header like so:

```
//go:build !ignore_autogenerated
// +build !ignore_autogenerated
```

And on Golang 1.17, gofmt will not complain and won't try to append
`go:build` and remove `+build`.

Signed-off-by: Sébastien Han <seb@redhat.com>