be able to run defaulter-gen tests via `go test` (#245)
* examples/defaulter-gen: be able to run tests via go test

Files whose names begin with "_" are ignored by go test so none of the
defaulter-gen tests were running in CI.

This commit renames `_output_tests` to `output_tests` so that tests are
run.

After doing so, tests failed because some packages missed a
`fake_deepcopy_conversion.go` file. This is necessary because types
registered in the scheme need to implement the `DeepCopyObject()` method.

Additionally, some of the tests import `k8s.io/api` and `k8s.io/apimachinery`
so these are also added to `go.mod`.

* move files back

* revert go.mod change

* revert makefile change

* add gomod and sum to _output_tests

* add comment about ambiguous import

* use &&

---------

Co-authored-by: Nikhita Raghunath <nikitaraghunath@gmail.com>