.*: add verify script for go directive changes
This commit adds a script that checks the changed version in
the go.mod file with a certain maximum version that the go
directive can have.

We set the maximum version of the go directive as 1.20 here
because the oldest go directive that exists on our supported
release branches in k/k is 1.20.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>