Avoid imports of same name
This primarily benefits applyconfiguration-gen today, which accidentaly
got this behavior by including the local package name in imports, which
caused goimports to deconflict with it before removing it since it was
unused.  This commit effictively preserves that behavior now that we
will not be relying on goimports to add/rename/remove imports.

Co-authored-by: Tim Hockin <thockin@google.com>