v2/parser: prefer original comment when evaluating aliased Type
Fixes the comment parsing on a Type so that it does not
get overwritten by an aliased Type with a different comment.
Depending on whether gotypealias=1|0, isTypeAlias will evaluate
whether an aliased type is detected as an alias in the Go type
system. When mismatched comments are detected on a type and its
alias, a warning is logged so that the API can be fixed to avoid
doing so.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>